requirements.txt
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.2
opencv-python>=4.8.0
mediapipe>=0.10.7
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
Pillow>=10.0.0
tqdm>=4.66.0
cachetools>=5.3.0
psutil>=5.9.5
transformers>=4.35.0
vaderSentiment>=3.3.2
newspaper3k>=0.2.8